Small refactoring. Move labId check and add tests.
