+++ Corona-Warn-App/src/test/java/de/rki/coronawarnapp/coronatest/CoronaTestTestComponent.kt
@@ -3,8 +3,8 @@ package de.rki.coronawarnapp.coronatest
 import dagger.Component
 import dagger.Module
 import de.rki.coronawarnapp.coronatest.type.TestCertificateContainerTest
-import de.rki.coronawarnapp.covidcertificate.test.TestCertificateRepositoryTest
-import de.rki.coronawarnapp.covidcertificate.test.storage.TestCertificateStorageTest
+import de.rki.coronawarnapp.covidcertificate.test.core.TestCertificateRepositoryTest
+import de.rki.coronawarnapp.covidcertificate.test.core.storage.TestCertificateStorageTest
 import de.rki.coronawarnapp.util.serialization.SerializationModule
 import javax.inject.Singleton
