+++ Corona-Warn-App/src/test/java/de/rki/coronawarnapp/covidcertificate/test/core/execution/TestCertificateRetrievalSchedulerTest.kt
@@ -1,4 +1,4 @@
-package de.rki.coronawarnapp.covidcertificate.test.execution
+package de.rki.coronawarnapp.covidcertificate.test.core.execution
 
 import androidx.work.OneTimeWorkRequest
 import androidx.work.WorkInfo
@@ -7,7 +7,6 @@ import de.rki.coronawarnapp.coronatest.CoronaTestRepository
 import de.rki.coronawarnapp.coronatest.type.CoronaTest
 import de.rki.coronawarnapp.covidcertificate.test.core.TestCertificateRepository
 import de.rki.coronawarnapp.covidcertificate.test.core.TestCertificateWrapper
-import de.rki.coronawarnapp.covidcertificate.test.core.execution.TestCertificateRetrievalScheduler
 import de.rki.coronawarnapp.util.device.ForegroundState
 import io.mockk.MockKAnnotations
 import io.mockk.Runs
