+++ Corona-Warn-App/src/test/java/de/rki/coronawarnapp/covidcertificate/test/core/storage/PCRTestCertificateTest.kt
@@ -1,4 +1,4 @@
-package de.rki.coronawarnapp.covidcertificate.test.storage
+package de.rki.coronawarnapp.covidcertificate.test.core.storage
 
 import testhelpers.BaseTest
