+++ Corona-Warn-App/src/test/java/de/rki/coronawarnapp/covidcertificate/test/core/storage/TestCertificateStorageTest.kt
@@ -1,10 +1,9 @@
-package de.rki.coronawarnapp.covidcertificate.test.storage
+package de.rki.coronawarnapp.covidcertificate.test.core.storage
 
 import android.content.Context
 import androidx.core.content.edit
 import de.rki.coronawarnapp.coronatest.DaggerCoronaTestTestComponent
 import de.rki.coronawarnapp.covidcertificate.test.TestCertificateTestData
-import de.rki.coronawarnapp.covidcertificate.test.core.storage.TestCertificateStorage
 import de.rki.coronawarnapp.covidcertificate.vaccination.core.repository.storage.ContainerPostProcessor
 import de.rki.coronawarnapp.util.serialization.SerializationModule
 import io.kotest.matchers.shouldBe
